Make null list+pushmark happen in more cases
This optimisation, added in 7d3c8a683, nulls the list+pushmark pair if
it occurs in list context, since the two ops effectively cancel each
other out.  I recently extended it in 8717a761e to apply to void con-
text, too.

It works by checking in the peephole optimiser whether the sibling of
the current op is a list with a pushmark kid:

1   current op
4   list
2     pushmark
3     ...
5   ...

That means the optimisation doesn’t happen if the elder sibling of the
list op is something that is not part of the execution chain, such as
$package_var:

-   ex-rv2sv
1     gvsv
4   list
2     pushmark
3     ...
5   ...

because the ex-rv2sv is never the ‘current op’.  So ($_,($_,$_))
doesn’t get optimised.

We can’t just handle this when ‘pushmark’ or ‘list’ is the current
op, because, in the former case, there is no way to get to the parent
op to null it; in the latter case, there is no way to get to the op
pointing to pushmark, to rethread the op_next pointers.

However, handling this when list or void context is applied, before we
even get to the peephole optimiser, just works, and requires much less
code.  We can just null the ops there, and leave it to the peephole
optimiser’s handling of null ops to rethread op_next pointers.

This breaks this convention:

    op_prepend_elem(OP_LIST, foo, list(that_list))

by creating two redundant null ops that were not there before, but the
only pieces of code doing that were calling list() needlessly anyway.