regcomp.c: study_chunk(): Avoid unnecessary recursion.
Avoid unnecessary recursion with regex subroutines. When we are inside
of a DEFINE we do not need to, and should not, recurse. And once an
infinite construct has been seen and we are not tracking the start
class then we do not /need/ to, at the cost of possibly pessimizing
minlen checks. For instance in

 /(?(DEFINE)(?<p>foo|wozzit))a+(?&p)/

entering 'p' via recursion would only tell us that the minlen was
actually 4 instead of being 1, something that would not make us
match incorrectly. This can be improved in a later commit.

The side effect of not recursing in such situations is that large
patterns such as that in RT #122283 compile in a reasonable amount
of time.