Remove very obsolete comment
I chuckle every time I read this comment.  And I chuckled at the time
it got ignored and mangled, too.  :-)

17ab7946 reworked some code without reindenting a block.  And that
lack of indentation was intentional and noted by this comment.

96913b52 four years later reindented the code, ignoring the comment.

486ec47a corrected the spelling mistake in the comment, rendering it
incomprehensible.  ‘Intenting’ was supposed to be ‘indenting’ (it was
obvious to me), but got ‘corrected’ to ‘intending’.

d5524600 gutted almost the entire block, moving it into a static rou-
tine, so the comment is by now not even remotely relevant.