add t/perf/speed.t
This test file is similar to /re/speed.t, but to test general-purpose
optimisations.

The idea is to run snippets of code that are 100s or 1000s times slower
if a particular optimisation is broken. We are not so much interested
in the individual tests passing, as in the whole file failing with a
watchdog timeout (or just observing that it running more slowly)