Support for nocapture regexp flag /n
This flag will prevent () from capturing and filling in $1, $2, etc...

Named captures will still work though, and if used will cause $1, $2, etc...
to be filled in *only* within named groups.

The motivation behind this is to allow the common construct of:

  /(?:b|c)a(?:t|n)/

To be rewritten more cleanly as:

  /(b|c)a(t|n)/n

When you want grouping but no memory penalty on captures.

You can also use ?n inside of a () directly to avoid capturing, and
?-n inside of a () to negate its effects if you want to capture.