Unixify the new Test::Simple temp dir on VMS.
We're going to be combining the temporary directory with hard-coded
forward slashes in various path components, so make sure the thing
we're appending to is also in Unix format.

Without this the core test suite on a threaded build hangs in
Thread::Queue because the Test::Simple bug causes lots of things to
fail and Thread::Queue doesn't handle threads that fail.

Hopefully submitted upstream at

<https://github.com/Test-More/test-more/issues/468>

but I have no idea what the relationship is between Test::More
and Test::Simple -- I just followed the link from search.cpan.org
and tried to open an "issue."