op.c: Rewrite op_const_sv
Originally, a compiled CV was checked to see whether it could be
inlined as a constant.  sub_const, later named, cv_const_sv, followed
op_next pointers starting at CvSTART to look for a single constant,
possibly surrounded by multiple nextstate, pushmark or null ops, fol-
lowed by leavesub or the end of the op chain.  Later this loop was
put in a separate function, op_const_sv, which is why it is laid out
like this.

It was 64d1d4c7d or fe5e78ed in 1998, whose purpose was to check the
values of constants before emitting a redefinition warning, that made
op_const_sv a separate function and caused it to be called before the
op_next pointers are set up.  That’s when things stopped making sense.

If you follow op_next pointers when they are not set up yet, you don’t
get anywhere.  And maybe that’s the point.  op_const_sv starts at the
second kid of the op passed to it, if the latter is a lineseq op.  So
a lone constant looks like this:

  lineseq
    nextstate
    const <-- op_const_sv starts here and looks no further

It never gets beyond that second kid.  In the case of two constants
separated by a semicolon, the first one becomes a null op, and the
surviving constant is never seen:

  lineseq
    nextstate
    null <-- op_const_sv starts here and looks no further
    nextstate
    const

It was that change that stopped sub(){return 3} and sub(){die if 0;5}
from being inlined (and sub(){return $outer_lexical}, for that matter,
which was inlined in perl 5.004¹), accidentally I presume, but the
altered behaviour has been documented now for over a decade, so it may
not be wise to revert it.

This commit just rewrites op_const_sv to work as currently, but with-
out the useless op_next loop.  In other works, it actually does now
what it looks as though it is doing, instead of requiring a day of
detective work to figure out why it works the way it does.

¹ I was not able to test this; I just surmised it based on the code.