lib/locale.t: Add knowledge about known bad locales
Some platforms currently have locale tests that always fail due to
platform-specific problems.  For example, we know that some Lithuanian
locales fail on Darwin because of a bad symbolic link there.

This commit adds the capability for locale.t to know about such issues
and not fail because of them.