[perl #121827] Fix repeat stack bugs
(...)x... is marked as a list at compile time and given a pushmark,
before the context is known.  If it turns out to be called in scalar
or void context after all, then pp_repeat has to handle the mark that
has been pushed on to the markstack.

It was not handling the mark for overloading.  Nor was it handling
void context correctly at all.  (The stack may have an empty list, in
which case we call FETCH on a potentially tied stack item that has
nothing to do with us.)

I tested it in void context, because I plan to undo the listification
when scalar context is applied, for speed.