op.c: Skip pushmark for (...)x... in scalar cx
(...)x... is marked as a list at compile time and given a pushmark,
before the context is known.  If it turns out to be called in scalar
or void context after all, then pp_repeat has to handle the mark that
has been pushed on to the markstack.

When the scalar context is applied, there is no reason not to undo
the listification, as long as there is only one left-hand argument.
That way we don’t have to execute a pushmark op and do the extra stack
fiddling at run time.