Allow OPpTARGET_MY optimisation for split
Many operators have a special SV allocated in the pad which is used
for return values (the target).  If we make that pad offset point to
a lexical variable, then we can optimise, say, $lexical = "$foo" into
just "$foo", where $lexical is stringify’s ‘target’.  And pp_stringify
doesn’t need to know any better.  We already do that for many ops.

This is safe to extend to split.  split only uses its target in this
code at the end:

    GETTARGET;
    PUSHi(iters);

so there is no danger of modifying its argument before reading it.