Remove OA_DANGEROUS from subst(cont)
OA_DANGEROUS indicates that temporary copies may need to be made in
list assignment, to handle things like:

($a, $b) = ($b, $a);

In other words, an op type is flagged with OA_DANGEROUS if its return
values could occur elsewhere on the stack.

pp_subst and pp_substcont only push new mortal scalars or read-only
values on to the stack.  Hence they are not OA_DANGEROUS.